Finished adding a delete form, search bar, and added routes
