+++ views/article.hbs
@@ -1,6 +1,6 @@
 <h1> {{title}} </h1>
+<br> <div>Author: {{author}}</div>
 <br> <div>{{body}}</div>
-<br> <div>{{author}}</div>
 <br>
 <br> <a href="/articles/{{title}}/edit">EDIT ITEM</a>
 <br> <a href="/articles/{{title}}/delete">DELETE ITEM</a>
