+++ views/index-redirect.hbs
@@ -10,7 +10,7 @@
 <h1>Products</h1>
 <br><form id="searchForm" onsubmit="submissionFunction()" method="get">
   <label> Search Product by Name:
-    <input type="text" name="name" id="productName">
+    <input type="text" name="name" id="productName" value="Name does not exist">
   </label>
 
   <input type="submit" value="Submit">
@@ -36,7 +36,7 @@
 <h1>Articles</h1>
 <br><form id="searchForm" onsubmit="submissionFunction()" method="get">
   <label> Search Article by Title:
-    <input type="text" name="name" id="articleTitle">
+    <input type="text" name="name" id="articleTitle" value="Title does not exist">
   </label>
 
   <input type="submit" value="Submit">
