+++ views/indexDeleted.hbs
@@ -2,8 +2,9 @@
 <h1>Products</h1>
 <br><a href="/products/new">NEW PRODUCT</a>
 <br><a href="/products/edit">EDIT PRODUCT</a>
+<br><a href="/products/delete">DELETE PRODUCT</a>
 <br>
-<br><div>Product:{{name}} has been deleted</div>
+<br><div>Product has been deleted</div>
 <br>
 <br> {{#each products}}
 <div>
@@ -15,8 +16,9 @@
 <h1>Articles</h1>
 <br><a href="/articles/new">NEW ARTICLE</a>
 <br><a href="/articles/edit">EDIT ARTICLE</a>
+<br><a href="/articles/delete">DELETE ARTICLE</a>
 <br>
-<br><div>Article:{{title}} has been deleted</div>
+<br><div>Article has been deleted</div>
 <br>
 <br> {{#each articles}}
 <div>
@@ -28,11 +30,11 @@
 <h1>Products</h1>
 <a href="/products/new">NEW PRODUCT</a>
 <br>
-<br><div>Product:{{name}} has been deleted</div>
+<br><div>Product has been deleted</div>
 <br>{{/with}} {{#with Article}}
 <h1>Articles</h1>
 <a href="/articles/new">NEW ARTICLE</a>
 <br>
-<br><div>Article:{{title}} has been deleted</div>
+<br><div>Article has been deleted</div>
 <br>{{/with}}
 <br> {{> footer}}
