+++ views/partials/footer.hbs
@@ -1,0 +1,3 @@
+<br>
+<br>
+<a href="/">HOME</a>
