+++ views/product.hbs
@@ -1,6 +1,6 @@
 <h1> {{name}} </h1>
-<br> <div>{{price}}</div>
-<br> <div>{{inventory}}</div>
+<br> <div>Price: {{price}}</div>
+<br> <div>Inventory: {{inventory}}</div>
 <br>
 <br> <a href="/products/{{id}}/edit">EDIT ITEM</a>
 <br> <a href="/products/{{id}}/delete">DELETE ITEM</a>
